Tidying Up The Project
